add filmsListController
